fix: dto 스펙 변경으로 인한 데이터 유형 변경

+++ app/src/main/java/kkj/mjc/ramenlog/like/LikeItem.java
@@ -4,9 +4,9 @@ public class LikeItem {
     int id;
     String name;
     String address;
-    int score;
+    double score;
 
-    public LikeItem(int id, String name, String address, int score) {
+    public LikeItem(int id, String name, String address, double score) {
         this.id = id;
         this.name = name;
         this.address = address;
